Bug fix: Conference Delete Bug Fix (#36)
* Fix delete_agenda_by_conference_id to handle
NoneType db_agenda

* Fix email sending error in otp_generator.py

* Fix email sending bug in OTP generator and handle
error in OTP router

* Fix SMTP connection and login errors in
otp_generator.py

* Fix email sending errors in otp_generator.py

* Fix SMTP server and port validation in
otp_generator.py

* Fix email server and port configuration

* Fix conditional statement in send_mail function

* Fix bug in delete_conference function

+++ app/crud/agenda_crud.py
@@ -127,7 +127,10 @@ def delete_agenda(db: Session, conference_id: str, attendee_id: str):
 # delete agenda by conference id
 def delete_agenda_by_conference_id(db: Session, conference_id: int):
     db_agenda = db.query(models.Agenda).filter(models.Agenda.conference_id == conference_id).first()
-    db.query(models.AgendaSession).filter(models.AgendaSession.agenda_id == db_agenda.id).delete()
-    db.delete(db_agenda)
-    db.commit()
-    return True
+    if db_agenda is not None:
+        db.query(models.AgendaSession).filter(models.AgendaSession.agenda_id == db_agenda.id).delete()
+        db.delete(db_agenda)
+        db.commit()
+        return True
+    else:
+        return False
